@@ -6,6 +6,8 @@
   "homepage": "https://cräft.com/",
   "scripts": {
     "clean": "rm -rf ./build/",
+    "prettier": "prettier ./src --write",
+    "lint": "eslint src --ext .js,.jsx,.ts,.tsx",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "build:itchio": "PUBLIC_URL=. REACT_APP_TARGET=itchio react-scripts build",
@@ -17,22 +19,30 @@
   "author": "Umlaut Games",
   "license": "MIT",
   "devDependencies": {
-    "create-react-app": "^3.3.0",
-    "eslint": "^6.8.0",
-    "eslint-plugin-flowtype": "^4.6.0",
-    "eslint-plugin-react": "^7.17.0",
-    "gh-pages": "^2.1.1",
-    "node-sass": "^7.0.0",
-    "prop-types": "^15.7.2",
-    "typescript": "^3.7.5"
+    "@babel/core": "^7.28.0",
+    "@babel/plugin-syntax-flow": "^7.27.1",
+    "@babel/plugin-transform-react-jsx": "^7.27.1",
+    "@eslint/compat": "^1.3.2",
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/jest-dom": "^6.6.4",
+    "@testing-library/react": "^16.3.0",
+    "create-react-app": "^5.1.0",
+    "eslint": "^9.33.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-jest": "^29.0.1",
+    "gh-pages": "^6.3.0",
+    "prettier": "^3.6.2",
+    "prop-types": "^15.8.1",
+    "sass": "^1.90.0",
+    "typescript": "^5.9.2"
   },
   "dependencies": {
-    "@craeft/engine": "git+https://github.com/craeftgame/engine.git",
-    "bulma": "^0.8.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-scripts": "^3.2.0",
-    "react-tooltip": "^4.0.3"
+    "@craeft/engine": "git+https://github.com/craeftgame/engine.git#v1.6.1",
+    "bulma": "^1.0.4",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-scripts": "^5.0.1",
+    "react-tooltip": "^5.29.1"
   },
   "browserslist": [
     ">0.2%",
@@ -43,5 +53,6 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/craeftgame/craeft.git"
-  }
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
