@@ -1,22 +1,19 @@
-import React, {Component} from "react";
+import React, { Component } from "react";
 
 import Header from "./components/structure/Header";
 import Footer from "./components/structure/Footer";
 import About from "./components/pages/About";
 
 export default class AboutPage extends Component {
-
     render() {
         return (
-            <div className='rpgui-content container'>
+            <div className="rpgui-content container">
+                <Header />
 
-                <Header/>
+                <About />
 
-                <About/>
-
-                <Footer/>
-
+                <Footer />
             </div>
-        )
+        );
     }
 }
