@@ -1,10 +1,9 @@
 /* globals it */
 import React from "react";
-import { createRoot } from "react-dom/client";
+import { render, screen } from "@testing-library/react";
 import CraeftPage from "../CraeftPage";
 
 it("renders without crashing", () => {
-    const root = createRoot(document.createElement("div"));
-    root.render(<CraeftPage showFooter={true} />);
-    root.unmount();
+    render(<CraeftPage showFooter={true} />);
+    expect(screen.getByText("Cräft!")).toBeInTheDocument();
 });
