@@ -1,8 +1,8 @@
 /* globals craeft */
-import React, {Component} from "react";
+import React, { Component } from "react";
 import PropTypes from "prop-types";
 import BossIcon from "./BossIcon";
-import BossDescription from "./BossDesctiption"
+import BossDescription from "./BossDesctiption";
 import Map from "../map/Map";
 import PlayerDescription from "../player/PlayerDescription";
 import Logs from "../Logs";
@@ -10,9 +10,8 @@ import Logs from "../Logs";
 import config from "@craeft/engine/config";
 
 export default class Adventure extends Component {
-
     static propTypes = {
-        onMove: PropTypes.func
+        onMove: PropTypes.func,
     };
 
     state = {
@@ -22,79 +21,87 @@ export default class Adventure extends Component {
     constructor(props) {
         super(props);
 
-        this.selectBoss = this.selectBoss.bind(this)
+        this.selectBoss = this.selectBoss.bind(this);
     }
 
     fight(id) {
         const boss = craeft.bosses.findById(id);
 
-        console.log(boss.name)
+        console.log(boss.name);
     }
 
     selectBoss(id) {
         this.setState({
-            selected: craeft.bosses.findById(id)
-        })
+            selected: craeft.bosses.findById(id),
+        });
     }
 
     render() {
         return (
-            <div className='adventure frame'>
-                <div className='rpgui-container framed'>
-
+            <div className="adventure frame">
+                <div className="rpgui-container framed">
                     <div className="row">
                         <strong>Adventure</strong>
-                        <hr/>
+                        <hr />
                     </div>
 
                     <div className="row">
                         <div className="columns">
-
                             <div className="column">
-                                <PlayerDescription player={craeft.player}/>
+                                <PlayerDescription player={craeft.player} />
 
-                                <Logs/>
+                                <Logs />
                             </div>
 
                             <div className="column">
-                                <Map onMove={this.props.onMove}/>
+                                <Map onMove={this.props.onMove} />
                             </div>
 
                             <div className="column">
+                                {config.showBossScreen ? (
+                                    <>
+                                        <div className="boss-name">
+                                            <span>
+                                                {this.state.selected.name}
+                                            </span>
+                                        </div>
 
-                                {
-                                    config.showBossScreen ?
-                                        <>
-                                            <div className="boss-name">
-                                                <span>{this.state.selected.name}</span>
-                                            </div>
+                                        <div>
+                                            {craeft.bosses.map((boss) => {
+                                                return (
+                                                    <BossIcon
+                                                        key={`boss-${boss.name}`}
+                                                        type={boss.type}
+                                                        isDead={boss.dead}
+                                                        isSelected={
+                                                            boss ===
+                                                            this.state.selected
+                                                        }
+                                                        onClick={() =>
+                                                            this.selectBoss(
+                                                                boss.id,
+                                                            )
+                                                        }
+                                                    />
+                                                );
+                                            })}
+                                        </div>
 
-                                            <div>
-
-                                                {
-                                                    craeft.bosses.map((boss) => {
-
-                                                        return <BossIcon key={`boss-${boss.name}`}
-                                                                         type={boss.type} isDead={boss.dead}
-                                                                         isSelected={boss === this.state.selected}
-                                                                         onClick={() => this.selectBoss(boss.id)}/>
-                                                    })
-                                                }
-
-                                            </div>
-
-                                            <BossDescription boss={this.state.selected}
-                                                             fight={() => this.fight(this.state.selected.id)}/>
-                                        </> : null
-                                }
-
+                                        <BossDescription
+                                            boss={this.state.selected}
+                                            fight={() =>
+                                                this.fight(
+                                                    this.state.selected.id,
+                                                )
+                                            }
+                                        />
+                                    </>
+                                ) : null}
                             </div>
-
                         </div>
                     </div>
-
                 </div>
             </div>
-        )
+        );
     }
 }
