@@ -1,11 +1,10 @@
-import React, {Component} from "react";
+import React, { Component } from "react";
 import Gauge from "../utility/Gauge";
-import {CraefterTypeNames} from "@craeft/engine/src/data/names";
+import { CraefterTypeNames } from "@craeft/engine/src/data/names";
 import PropTypes from "prop-types";
 import Attribute from "../utility/Attribute";
 
 export default class CraefterDescription extends Component {
-
     static propTypes = {
         craefter: PropTypes.object,
     };
@@ -13,59 +12,66 @@ export default class CraefterDescription extends Component {
     render() {
         return (
             <div>
-
                 <div className="row">
                     <span>{CraefterTypeNames[this.props.craefter.type]}</span>
                 </div>
 
-                {
-                    !this.props.craefter.delay.isDelaying ?
+                {!this.props.craefter.delay.isDelaying ? (
+                    <div className="row">
+                        Level {this.props.craefter.level}:{" "}
+                        <span>{this.props.craefter.name}</span>
+                    </div>
+                ) : (
+                    "???"
+                )}
+
+                {!this.props.craefter.delay.isDelaying ? (
+                    <div>
                         <div className="row">
-                            Level {this.props.craefter.level}: <span>{this.props.craefter.name}</span>
-                        </div>
-                        : "???"
-                }
-
-                {
-                    !this.props.craefter.delay.isDelaying ?
-                        <div>
-
-                            <div className='row'>
-
-                                <div>
-                                    <Attribute label="Str" value={this.props.craefter.str}/>
-                                    <Attribute label="Dex" value={this.props.craefter.dex}/>
-                                </div>
-
-                                <div>
-                                    <Attribute label="Int" value={this.props.craefter.int}/>
-                                    <Attribute label="Luk" value={this.props.craefter.luk}/>
-                                </div>
-
+                            <div>
+                                <Attribute
+                                    label="Str"
+                                    value={this.props.craefter.str}
+                                />
+                                <Attribute
+                                    label="Dex"
+                                    value={this.props.craefter.dex}
+                                />
                             </div>
 
-                            <hr/>
-
-                            <div className='row'>
-
-                                <Gauge label='STA' color='green'
-                                       current={this.props.craefter.staCurrent}
-                                       max={this.props.craefter.staMax}/>
-
-
-                                <Gauge label='EXP'
-                                       current={this.props.craefter.expCurrent}
-                                       max={this.props.craefter.expMax}/>
-
+                            <div>
+                                <Attribute
+                                    label="Int"
+                                    value={this.props.craefter.int}
+                                />
+                                <Attribute
+                                    label="Luk"
+                                    value={this.props.craefter.luk}
+                                />
                             </div>
-
-                            <hr/>
-
                         </div>
-                        : null
-                }
 
+                        <hr />
+
+                        <div className="row">
+                            <Gauge
+                                label="STA"
+                                color="green"
+                                current={this.props.craefter.staCurrent}
+                                max={this.props.craefter.staMax}
+                            />
+
+                            <Gauge
+                                label="EXP"
+                                current={this.props.craefter.expCurrent}
+                                max={this.props.craefter.expMax}
+                            />
+                        </div>
+
+                        <hr />
+                    </div>
+                ) : null}
             </div>
-        )
+        );
     }
 }
