@@ -1,21 +1,20 @@
-import React, {Component} from "react";
+import React, { Component } from "react";
 import PropTypes from "prop-types";
 
 import CraeftingWindow from "../dialogs/CraeftingWindow";
 import AddCraeftersWindow from "../dialogs/AddCraeftersWindow";
 import Craefter from "./Craefter";
 import CraefterIcon from "./CraefterIcon";
-import {ResourceTypes} from "@craeft/engine/src/data/types";
+import { ResourceTypes } from "@craeft/engine/src/data/types";
 import Modal from "../utility/Modal";
 
 export default class CraefterList extends Component {
-
     static propTypes = {
         craefters: PropTypes.array,
         resources: PropTypes.object,
         itemAdded: PropTypes.func,
         craefterAdded: PropTypes.func,
-        bury: PropTypes.func
+        bury: PropTypes.func,
     };
 
     state = {
@@ -24,7 +23,7 @@ export default class CraefterList extends Component {
         // is the add craefter dialog showing?
         isAddCraefterDialogShown: false,
         // currently selected craefter
-        selectedCraefter: null
+        selectedCraefter: null,
     };
 
     constructor(props) {
@@ -53,45 +52,38 @@ export default class CraefterList extends Component {
 
     openCraefterDialog() {
         this.setState({
-            isAddCraefterDialogShown: true
-        })
+            isAddCraefterDialogShown: true,
+        });
     }
 
     openCraeftDialog(craefter) {
         this.setState({
             isCraeftingDialogShown: true,
-            selectedCraefter: craefter
+            selectedCraefter: craefter,
         });
     }
 
-    addCraefter(
-        craefter
-    ) {
+    addCraefter(craefter) {
         this.props.craefterAdded(craefter);
         this.setState({
-            isAddCraefterDialogShown: false
-        })
+            isAddCraefterDialogShown: false,
+        });
     }
 
-    addItem(
-        item,
-        resourcesConsumed
-    ) {
+    addItem(item, resourcesConsumed) {
         this.props.itemAdded(
             item,
             resourcesConsumed,
-            this.state.selectedCraefter
+            this.state.selectedCraefter,
         );
 
         this.setState({
             isCraeftingDialogShown: false,
-            selectedCraefter: null
-        })
+            selectedCraefter: null,
+        });
     }
 
-    onCraefterSelect(
-        craefter
-    ) {
+    onCraefterSelect(craefter) {
         let c = this.state.selectedCraefter === craefter ? null : craefter;
 
         if (craefter.isCraefting) {
@@ -99,92 +91,97 @@ export default class CraefterList extends Component {
         }
 
         this.setState({
-            selectedCraefter: c
-        })
+            selectedCraefter: c,
+        });
     }
 
-    bury(
-        craefterId
-    ) {
+    bury(craefterId) {
         this.props.bury(craefterId);
 
         this.setState({
-            selectedCraefter: null
-        })
+            selectedCraefter: null,
+        });
     }
 
     render() {
         return (
-            <div className='craefters column frame'>
+            <div className="craefters column frame">
+                <div className="rpgui-container framed">
+                    {this.state.isCraeftingDialogShown ? (
+                        <Modal
+                            isActive={this.state.isCraeftingDialogShown}
+                            onClose={() =>
+                                this.setState({
+                                    isCraeftingDialogShown: false,
+                                })
+                            }
+                        >
+                            <CraeftingWindow
+                                resources={this.props.resources}
+                                craefter={this.state.selectedCraefter}
+                                itemAdded={this.addItem}
+                            />
+                        </Modal>
+                    ) : null}
 
-                <div className='rpgui-container framed'>
-
-                    {
-                        this.state.isCraeftingDialogShown ?
-                            <Modal isActive={this.state.isCraeftingDialogShown}
-                                   onClose={() => this.setState({
-                                       isCraeftingDialogShown: false
-                                   })}>
-                                <CraeftingWindow resources={this.props.resources}
-                                                 craefter={this.state.selectedCraefter}
-                                                 itemAdded={this.addItem}/>
-                            </Modal> : null
-                    }
-
-                    {
-                        this.state.isAddCraefterDialogShown ?
-                            <Modal isActive={this.state.isAddCraefterDialogShown}
-                                   onClose={() => this.setState({
-                                       isAddCraefterDialogShown: false
-                                   })}>
-                                <AddCraeftersWindow addCraefter={this.addCraefter}/>
-                            </Modal> : null
-                    }
+                    {this.state.isAddCraefterDialogShown ? (
+                        <Modal
+                            isActive={this.state.isAddCraefterDialogShown}
+                            onClose={() =>
+                                this.setState({
+                                    isAddCraefterDialogShown: false,
+                                })
+                            }
+                        >
+                            <AddCraeftersWindow
+                                addCraefter={this.addCraefter}
+                            />
+                        </Modal>
+                    ) : null}
 
                     <div className={"row"}>
                         <strong>Cräfter</strong>
-                        <hr/>
+                        <hr />
                     </div>
 
-                    {
-                        this.state.selectedCraefter && !this.state.selectedCraefter.isCraefting ?
-                            <Craefter craefter={this.state.selectedCraefter}
-                                      openCraeftDialog={this.openCraeftDialog}
-                                      bury={this.bury}
-                                      canCraeft={this.hasEnoughResources()}/>
-                            : null
-                    }
+                    {this.state.selectedCraefter &&
+                    !this.state.selectedCraefter.isCraefting ? (
+                        <Craefter
+                            craefter={this.state.selectedCraefter}
+                            openCraeftDialog={this.openCraeftDialog}
+                            bury={this.bury}
+                            canCraeft={this.hasEnoughResources()}
+                        />
+                    ) : null}
 
-                    {
-                        this.props.craefters.map((craefter, index) => {
-                            return (
-                                <CraefterIcon key={index} craefter={craefter}
-                                              isSelected={
-                                                  this.state.selectedCraefter === craefter &&
-                                                  !this.state.selectedCraefter.isCraefting
-                                              }
-                                              onCraefterSelect={this.onCraefterSelect}/>
-                            )
-                        })
-                    }
+                    {this.props.craefters.map((craefter, index) => {
+                        return (
+                            <CraefterIcon
+                                key={index}
+                                craefter={craefter}
+                                isSelected={
+                                    this.state.selectedCraefter === craefter &&
+                                    !this.state.selectedCraefter.isCraefting
+                                }
+                                onCraefterSelect={this.onCraefterSelect}
+                            />
+                        );
+                    })}
 
                     <div className={"row"}>
-                        <button onClick={this.openCraefterDialog}
-                                className='rpgui-button is-small'>
-
+                        <button
+                            onClick={this.openCraefterDialog}
+                            className="rpgui-button is-small"
+                        >
                             <span className="icon">
-                                <i className="fas fa-plus"/>
+                                <i className="fas fa-plus" />
                             </span>
 
-                            <span>
-                                &nbsp;Add Cräfter
-                            </span>
-
+                            <span>&nbsp;Add Cräfter</span>
                         </button>
                     </div>
-
                 </div>
             </div>
-        )
+        );
     }
 }
