@@ -1,22 +1,18 @@
-import React, {Component} from "react";
+import React, { Component } from "react";
 import PropTypes from "prop-types";
 
 import ItemDescription from "../item/ItemDescription";
 import ItemIcon from "../item/ItemIcon";
-import {
-    ArmorSlots,
-    WeaponSlots
-} from "@craeft/engine/src/data/types";
+import { ArmorSlots, WeaponSlots } from "@craeft/engine/src/data/types";
 
 export default class Equipment extends Component {
-
     static propTypes = {
         onUnequip: PropTypes.func,
-        equipment: PropTypes.object
+        equipment: PropTypes.object,
     };
 
     state = {
-        selectedItem: null
+        selectedItem: null,
     };
 
     constructor(props) {
@@ -26,17 +22,13 @@ export default class Equipment extends Component {
         this.unequip = this.unequip.bind(this);
     }
 
-    toggleItemSelection(
-        item
-    ) {
+    toggleItemSelection(item) {
         this.setState({
-            selectedItem: this.state.selectedItem === item ? null : item
-        })
+            selectedItem: this.state.selectedItem === item ? null : item,
+        });
     }
 
-    unequip(
-        item
-    ) {
+    unequip(item) {
         this.props.onUnequip(item.id);
 
         if (!item.equipped) {
@@ -46,110 +38,153 @@ export default class Equipment extends Component {
 
     render() {
         return (
-            <div className='rpgui-container framed-grey row equipment'>
-
+            <div className="rpgui-container framed-grey row equipment">
                 <div>
                     <strong>Equipment</strong>
                 </div>
 
-                <hr/>
+                <hr />
 
-                <div className='columns'>
-                    <div className='column'>
-                        Head<br/>
-                        {
-                            this.props.equipment[ArmorSlots.Head] ?
-                                <ItemIcon item={this.props.equipment[ArmorSlots.Head]}
-                                          isSelected={this.props.equipment[ArmorSlots.Head] === this.state.selectedItem}
-                                          isSmall={true} onItemSelected={this.toggleItemSelection}/>
-                                : <span className="rpgui-icon armor-slot"/>
-                        }
+                <div className="columns">
+                    <div className="column">
+                        Head
+                        <br />
+                        {this.props.equipment[ArmorSlots.Head] ? (
+                            <ItemIcon
+                                item={this.props.equipment[ArmorSlots.Head]}
+                                isSelected={
+                                    this.props.equipment[ArmorSlots.Head] ===
+                                    this.state.selectedItem
+                                }
+                                isSmall={true}
+                                onItemSelected={this.toggleItemSelection}
+                            />
+                        ) : (
+                            <span className="rpgui-icon armor-slot" />
+                        )}
                     </div>
-                    <div className='column'>
-                        Body<br/>
-                        {
-                            this.props.equipment[ArmorSlots.Body] ?
-                                <ItemIcon item={this.props.equipment[ArmorSlots.Body]}
-                                          isSelected={this.props.equipment[ArmorSlots.Body] === this.state.selectedItem}
-                                          isSmall={true} onItemSelected={this.toggleItemSelection}/>
-                                : <span className="rpgui-icon armor-slot"/>
-                        }
-                    </div>
-                </div>
-
-                <div className='columns'>
-                    <div className='column'>
-                        Legs<br/>
-                        {
-                            this.props.equipment[ArmorSlots.Legs] ?
-                                <ItemIcon item={this.props.equipment[ArmorSlots.Legs]}
-                                          isSelected={this.props.equipment[ArmorSlots.Legs] === this.state.selectedItem}
-                                          isSmall={true} onItemSelected={this.toggleItemSelection}/>
-                                : <span className="rpgui-icon armor-slot"/>
-                        }
-                    </div>
-                    <div className='column'>
-                        Feet<br/>
-                        {
-                            this.props.equipment[ArmorSlots.Feet] ?
-                                <ItemIcon item={this.props.equipment[ArmorSlots.Feet]}
-                                          isSelected={this.props.equipment[ArmorSlots.Feet] === this.state.selectedItem}
-                                          isSmall={true} onItemSelected={this.toggleItemSelection}/>
-                                : <span className="rpgui-icon armor-slot"/>
-                        }
+                    <div className="column">
+                        Body
+                        <br />
+                        {this.props.equipment[ArmorSlots.Body] ? (
+                            <ItemIcon
+                                item={this.props.equipment[ArmorSlots.Body]}
+                                isSelected={
+                                    this.props.equipment[ArmorSlots.Body] ===
+                                    this.state.selectedItem
+                                }
+                                isSmall={true}
+                                onItemSelected={this.toggleItemSelection}
+                            />
+                        ) : (
+                            <span className="rpgui-icon armor-slot" />
+                        )}
                     </div>
                 </div>
 
-                <div className='columns'>
-                    <div className='column'>
-                        Left Hand:<br/>
-                        {
-                            this.props.equipment[WeaponSlots.LeftHand] ?
-                                <ItemIcon item={this.props.equipment[WeaponSlots.LeftHand]}
-                                          isSelected={this.props.equipment[WeaponSlots.LeftHand] === this.state.selectedItem}
-                                          isSmall={true} onItemSelected={this.toggleItemSelection}/>
-                                : <span className="rpgui-icon weapon-slot"/>
-                        }
+                <div className="columns">
+                    <div className="column">
+                        Legs
+                        <br />
+                        {this.props.equipment[ArmorSlots.Legs] ? (
+                            <ItemIcon
+                                item={this.props.equipment[ArmorSlots.Legs]}
+                                isSelected={
+                                    this.props.equipment[ArmorSlots.Legs] ===
+                                    this.state.selectedItem
+                                }
+                                isSmall={true}
+                                onItemSelected={this.toggleItemSelection}
+                            />
+                        ) : (
+                            <span className="rpgui-icon armor-slot" />
+                        )}
                     </div>
-                    <div className='column'>
-                        Right Hand<br/>
-                        {
-                            this.props.equipment[WeaponSlots.RightHand] ?
-                                <ItemIcon item={this.props.equipment[WeaponSlots.RightHand]}
-                                          isSelected={this.props.equipment[WeaponSlots.RightHand] === this.state.selectedItem}
-                                          isSmall={true} onItemSelected={this.toggleItemSelection}/>
-                                : <span className="rpgui-icon weapon-slot"/>
-                        }
+                    <div className="column">
+                        Feet
+                        <br />
+                        {this.props.equipment[ArmorSlots.Feet] ? (
+                            <ItemIcon
+                                item={this.props.equipment[ArmorSlots.Feet]}
+                                isSelected={
+                                    this.props.equipment[ArmorSlots.Feet] ===
+                                    this.state.selectedItem
+                                }
+                                isSmall={true}
+                                onItemSelected={this.toggleItemSelection}
+                            />
+                        ) : (
+                            <span className="rpgui-icon armor-slot" />
+                        )}
                     </div>
                 </div>
 
-                <div className='columns'>
-
-                    <div className='column'>
-                        Jewelry 1<br/>
-                        <span className="rpgui-icon ring-slot"/>
+                <div className="columns">
+                    <div className="column">
+                        Left Hand:
+                        <br />
+                        {this.props.equipment[WeaponSlots.LeftHand] ? (
+                            <ItemIcon
+                                item={
+                                    this.props.equipment[WeaponSlots.LeftHand]
+                                }
+                                isSelected={
+                                    this.props.equipment[
+                                        WeaponSlots.LeftHand
+                                    ] === this.state.selectedItem
+                                }
+                                isSmall={true}
+                                onItemSelected={this.toggleItemSelection}
+                            />
+                        ) : (
+                            <span className="rpgui-icon weapon-slot" />
+                        )}
                     </div>
-
-                    <div className='column'>
-                        Jewelry 2<br/>
-                        <span className="rpgui-icon ring-slot"/>
+                    <div className="column">
+                        Right Hand
+                        <br />
+                        {this.props.equipment[WeaponSlots.RightHand] ? (
+                            <ItemIcon
+                                item={
+                                    this.props.equipment[WeaponSlots.RightHand]
+                                }
+                                isSelected={
+                                    this.props.equipment[
+                                        WeaponSlots.RightHand
+                                    ] === this.state.selectedItem
+                                }
+                                isSmall={true}
+                                onItemSelected={this.toggleItemSelection}
+                            />
+                        ) : (
+                            <span className="rpgui-icon weapon-slot" />
+                        )}
                     </div>
-
                 </div>
 
-                {
-                    this.state.selectedItem && this.state.selectedItem.equipped ?
-                        <div>
+                <div className="columns">
+                    <div className="column">
+                        Jewelry 1<br />
+                        <span className="rpgui-icon ring-slot" />
+                    </div>
 
-                            <hr/>
+                    <div className="column">
+                        Jewelry 2<br />
+                        <span className="rpgui-icon ring-slot" />
+                    </div>
+                </div>
 
-                            <ItemDescription item={this.state.selectedItem}
-                                             onUnequip={this.unequip}/>
+                {this.state.selectedItem && this.state.selectedItem.equipped ? (
+                    <div>
+                        <hr />
 
-                        </div> : null
-                }
-
+                        <ItemDescription
+                            item={this.state.selectedItem}
+                            onUnequip={this.unequip}
+                        />
+                    </div>
+                ) : null}
             </div>
-        )
+        );
     }
 }
