@@ -1,45 +1,51 @@
-import React, { Component } from 'react'
-import Donate from './Donate'
-import PropTypes from 'prop-types'
+import React, { Component } from "react";
+import Donate from "./Donate";
+import PropTypes from "prop-types";
 
 export default class Footer extends Component {
-
     static propTypes = {
-        showFooter: PropTypes.bool.isRequired
-    }
+        showFooter: PropTypes.bool.isRequired,
+    };
 
     render() {
         return (
             <div className="row footer">
+                <hr className="golden" />
 
-                <hr className="golden"/>
-
-                {
-                    this.props.showFooter ?
-                        <div className="columns">
-
-                            <div className="column">
-                                <Donate/>
-                            </div>
-
-                            <div className="column" style={{
-                                textAlign: 'left'
-                            }}>
-                                <a href={'#about'}>About Cräft!</a>
-                            </div>
-
+                {this.props.showFooter ? (
+                    <div className="columns">
+                        <div className="column">
+                            <Donate />
                         </div>
-                        : null
-                }
 
-                <div className="rpgui-center"
-                     style={{
-                         fontSize: '12px'
-                     }}>
-                    © 2019-2020 <a href="https://umlaut.games" target="_blank">Umlaut Games</a> — All Rights Reserved
+                        <div
+                            className="column"
+                            style={{
+                                textAlign: "left",
+                            }}
+                        >
+                            <a href={"#about"}>About Cräft!</a>
+                        </div>
+                    </div>
+                ) : null}
+
+                <div
+                    className="rpgui-center"
+                    style={{
+                        fontSize: "12px",
+                    }}
+                >
+                    © 2019-2025{" "}
+                    <a
+                        href="https://umlaut.games"
+                        target="_blank"
+                        rel="noreferrer"
+                    >
+                        Umlaut Games
+                    </a>{" "}
+                    — All Rights Reserved
                 </div>
-
             </div>
-        )
+        );
     }
 }
